savanah028-fixed
I changed forms.h to call tanks::move with wall array so I could see if
it works well and it does.

I have a slight glitch- I made it so if the tank collides with the wall
it moves away from the wall 5 pixels (so it doesnt get stuck) and it
works well except for if the tank is lined up perfectly with the edge of
the wall.... idk how to fix this. but its unlikely itll line up
perfectly anyway